Convert SCP HDF5 JS from web GUI to CLI
