--- hdf5-js/page.html
@@ -1,9 +1,0 @@
-<html>
-  <head>
-    <script src="https://cdn.jsdelivr.net/npm/jsfive@0.3.13/dist/browser/hdf5.js"></script>
-    <script src="lib.js"></script>
-  </head>
-  <body>
-    <input type="file" id="datafile" onchange="readHDF5()"/>
-  </body>
-</html>
